grove: Move lcd-grove into grove
