+++ Makefile.am
@@ -56,7 +56,6 @@ AM_CPPFLAGS = \
 	-I$(top_builddir)/src/modules/flow/hub \
 	-I$(top_builddir)/src/modules/flow/int \
 	-I$(top_builddir)/src/modules/flow/keyboard \
-	-I$(top_builddir)/src/modules/flow/lcd-grove \
 	-I$(top_builddir)/src/modules/flow/network \
 	-I$(top_builddir)/src/modules/flow/oic \
 	-I$(top_builddir)/src/modules/flow/piezo-speaker \
