+++ configure.ac
@@ -347,7 +347,6 @@ SOL_FLOW_MODULE([oic], [yes], [CHECK_FLOW_OIC])
 SOL_FLOW_MODULE([fs], [yes])
 SOL_FLOW_MODULE([test], [yes])
 SOL_FLOW_MODULE([servo-motor], [yes])
-SOL_FLOW_MODULE([lcd-grove], [yes])
 SOL_FLOW_MODULE([dummy], [no])
 SOL_FLOW_MODULE([network], [${platform_linux}])
 SOL_FLOW_MODULE([udev], [${platform_linux}], [CHECK_FLOW_UDEV])
