--- src/modules/flow/lcd-grove/lcd-grove.json
@@ -1,214 +1,0 @@
-{
-  "name": "lcd-grove",
-  "meta": {
-    "author": "Intel Corporation",
-    "license": "BSD 3-Clause",
-    "version": "1"
-  },
-  "types": [
-    {
-      "category": "output/hw",
-      "description": "LCD grove controller (JHD131M1 model) for simple string displaying and backlight color setting. It uses two I2C addresses, one to control an HD44780 based display and one for RGB backlight. Display matrix is 16x2.",
-      "in_ports": [
-        {
-          "data_type": "string",
-          "description": "Set string to be displayed. If it's bigger than 16 characters, it breaks it to the second line. If it's bigger than 32 (what fits on the two display lines), it'll be cropped to this length.",
-          "methods": {
-            "process": "set_string"
-          },
-          "name": "IN"
-        },
-        {
-          "data_type": "rgb",
-          "description": "Set display's backlight color.",
-          "methods": {
-            "process": "set_color"
-          },
-          "name": "COLOR"
-        }
-      ],
-      "methods": {
-        "close": "lcd_close",
-        "open": "lcd_string_open"
-      },
-      "name": "lcd-grove/string",
-      "private_data_type": "lcd_data",
-      "url": "http://solettaproject.org/doc/latest/components/lcd/grove.html"
-    },
-    {
-      "category": "output/hw",
-      "description": "LCD grove controller (JHD131M1 model) for custom character/cursor manipulation and backlight color setting. It uses two I2C addresses, one to control an HD44780 based display and one for RGB backlight. Display matrix is 16x2.",
-      "in_ports": [
-        {
-          "data_type": "boolean",
-          "description": "Set whether auto-scrolling is on or off. After it's set to on, the cursor position is held static until one sets it off again (or changes it explicitly).",
-          "methods": {
-            "process": "set_autoscroll"
-          },
-          "name": "AUTO_SCROLL"
-        },
-        {
-          "data_type": "boolean",
-          "description": "Set character entry direction to be left-to-right (true) or right-to-left (false).",
-          "methods": {
-            "process": "set_ltr"
-          },
-          "name": "LEFT_TO_RIGHT"
-        },
-        {
-          "data_type": "rgb",
-          "description": "Set display's backlight color.",
-          "methods": {
-            "process": "set_color"
-          },
-          "name": "COLOR"
-        },
-        {
-          "data_type": "int",
-          "description": "Set display's current column. Valid range is 0-16 (16 is a virtual position useful when writing on the last column in right-to-left mode).",
-          "methods": {
-            "process": "set_col"
-          },
-          "name": "COL"
-        },
-        {
-          "data_type": "int",
-          "description": "Set display's current row. Valid range is 0-1.",
-          "methods": {
-            "process": "set_row"
-          },
-          "name": "ROW"
-        },
-        {
-          "data_type": "boolean",
-          "description": "Set whether the display's blinking cursor is on (true) or off (false).",
-          "methods": {
-            "process": "set_blinking_cursor"
-          },
-          "name": "BLINK_CURSOR"
-        },
-        {
-          "data_type": "boolean",
-          "description": "Set whether the display's underline cursor is on (true) or off (false).",
-          "methods": {
-            "process": "set_underline_cursor"
-          },
-          "name": "UNDERLINE_CURSOR"
-        },
-        {
-          "data_type": "boolean",
-          "description": "Set whether the display is on (true) or not (false). Naturally, it starts on by default.",
-          "methods": {
-            "process": "set_display_on"
-          },
-          "name": "DISPLAY_ON"
-        },
-        {
-          "data_type": "byte",
-          "description": "Write given character at display's current row/column and advance them one position. The column will raise or lower depending on the entry direction, and no wrapping will occur -- if either top-left or bottom-right of the screen is hit, further char writes will update those limit cells until one changes the cursor position explicitly.",
-          "methods": {
-            "process": "put_char"
-          },
-          "name": "PUT_CHAR"
-        },
-        {
-          "data_type": "string",
-          "description": "Write a sequence of characters at display's current row/column and advance the cursor's position. The cursor's position will change differently depending on the entry direction, and, if either top-left or bottom-right of the screen is hit, the string will be cropped at that point.",
-          "methods": {
-            "process": "put_string"
-          },
-          "name": "PUT_STRING"
-        },
-        {
-          "data_type": "any",
-          "description": "Shift/scroll right the whole display by one column, keeping the char cell states (lcd's RAM state) as it was. Thus, a subsequent shift left will bring the display to what it looked like before the first shift.",
-          "methods": {
-            "process": "scroll_display"
-          },
-          "name": "SCROLL_RIGHT"
-        },
-        {
-          "data_type": "any",
-          "description": "Shift/scroll left the whole display by one column, keeping the char cell states (lcd's RAM state) as it was. Thus, a subsequent shift left will bring the display to what it looked like before the first shift.",
-          "methods": {
-            "process": "scroll_display"
-          },
-          "name": "SCROLL_LEFT"
-        },
-        {
-          "data_type": "any",
-          "description": "Clear the whole display, also resetting cursor to the top-left position.",
-          "methods": {
-            "process": "display_clear"
-          },
-          "name": "CLEAR"
-        }
-      ],
-      "methods": {
-        "close": "lcd_close",
-        "open": "lcd_char_open"
-      },
-      "name": "lcd-grove/char",
-      "options": {
-       "members": [
-        {
-          "data_type": "int",
-          "default": {
-            "val": 0
-          },
-          "description": "Initial column of the cursor.",
-          "name": "init_col"
-        },
-        {
-          "data_type": "int",
-          "default": {
-            "val": 0
-          },
-          "description": "Initial row of the cursor.",
-          "name": "init_row"
-        },
-        {
-          "data_type": "boolean",
-          "default": true,
-          "description": "Whether to start on left-to-right (true) or right-to-left (false) cursor advancing.",
-          "name": "ltr"
-        },
-        {
-          "data_type": "boolean",
-          "default": false,
-          "description": "Set whether the display's blinking cursor is on (true) or off (false).",
-          "name": "blink_cursor"
-        },
-        {
-          "data_type": "boolean",
-          "default": false,
-          "description": "Set whether the display's underline cursor is on (true) or off (false).",
-          "name": "underline_cursor"
-        },
-        {
-          "data_type": "boolean",
-          "default": false,
-          "description": "Whether to start on auto scroll mode (true) or not (false).",
-          "name": "auto_scroll"
-        },
-        {
-          "data_type": "rgb",
-          "default": {
-            "red": 255,
-            "green": 255,
-            "blue": 255,
-            "red_max": 255,
-            "green_max": 255,
-            "blue_max": 255
-          },
-          "description": "Initial background color.",
-          "name": "color"
-        }
-       ],
-       "version": 1
-       },
-      "private_data_type": "lcd_data",
-      "url": "http://solettaproject.org/doc/latest/components/lcd/grove.html"
-    }
-  ]
-}
