+++ src/samples/flow/galileo-grove-kit/galileo-grove-kit.conf
@@ -1,9 +1,9 @@
 #i2c (any i2c shield port will do)
 [SolettaNodeEntry LCDString]
-Type=lcd-grove/string
+Type=grove/lcd-string
 
 [SolettaNodeEntry LCDChar]
-Type=lcd-grove/char
+Type=grove/lcd-char
 
 #analog
 # ANL0
